+++ lib/src/widgets/checkbox.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_omarchy/src/omarchy.dart';
+import 'package:flutter_omarchy/flutter_omarchy.dart';
 import 'package:flutter_omarchy/src/theme/colors.dart';
+import 'package:flutter_omarchy/src/theme/theme.dart';
 import 'package:flutter_omarchy/src/widgets/icon_data.g.dart';
 
 enum CheckboxState { checked, unchecked, multiple }
@@ -12,6 +13,7 @@ class OmarchyCheckbox extends StatelessWidget {
     bool isChecked = false,
     this.accent,
     this.size,
+    this.onPressed,
   }) : state =
            state ??
            (isChecked ? CheckboxState.checked : CheckboxState.unchecked);
@@ -19,10 +21,11 @@ class OmarchyCheckbox extends StatelessWidget {
   final CheckboxState state;
   final AnsiColor? accent;
   final double? size;
+  final VoidCallback? onPressed;
 
   @override
   Widget build(BuildContext context) {
-    final theme = Omarchy.of(context).theme;
+    final theme = OmarchyTheme.of(context);
     final foreground = accent != null
         ? theme.colors.bright[accent!]
         : theme.colors.foreground;
@@ -34,31 +37,51 @@ class OmarchyCheckbox extends StatelessWidget {
       CheckboxState.checked || CheckboxState.multiple => true,
     };
     final size = this.size ?? (theme.text.normal.fontSize ?? 12) * 1.2;
-    return AnimatedContainer(
-      duration: const Duration(milliseconds: 120),
-      width: size,
-      height: size,
-      decoration: BoxDecoration(
-        color: isSelected ? background : background.withValues(alpha: 0),
-        border: Border.all(
-          color: isSelected ? foreground : theme.colors.normal.white,
-          width: 2,
-        ),
-      ),
-      child: AnimatedSwitcher(
-        duration: const Duration(milliseconds: 120),
-        child: IconTheme(
-          data: IconThemeData(color: foreground, size: size * 0.6),
-          child: switch (state) {
-            CheckboxState.checked => const Icon(OmarchyIcons.codCheck),
-            CheckboxState.unchecked => const SizedBox.shrink(),
-            CheckboxState.multiple => Container(
-              color: foreground,
-              margin: const EdgeInsets.all(4.0),
+    return PointerArea(
+      onTap: onPressed,
+      builder: (context, state, child) {
+        final border = switch (state) {
+          PointerState(isHovering: true) when isSelected =>
+            foreground.withValues(alpha: foreground.a * 0.9),
+          PointerState(isHovering: true) => foreground.withValues(
+            alpha: foreground.a * 0.6,
+          ),
+          _ when isSelected => foreground,
+          _ => theme.colors.normal.white,
+        };
+        final fill = switch (state) {
+          PointerState(isHovering: true) when isSelected =>
+            background.withValues(alpha: background.a * 0.6),
+          PointerState(isHovering: true) => background.withValues(
+            alpha: background.a * 0.6,
+          ),
+          _ when isSelected => background,
+          _ => theme.colors.normal.black,
+        };
+        return AnimatedContainer(
+          duration: const Duration(milliseconds: 120),
+          width: size,
+          height: size,
+          decoration: BoxDecoration(
+            color: fill,
+            border: Border.all(color: border, width: 2),
+          ),
+          child: AnimatedSwitcher(
+            duration: const Duration(milliseconds: 120),
+            child: IconTheme(
+              data: IconThemeData(color: foreground, size: size * 0.6),
+              child: switch (this.state) {
+                CheckboxState.checked => const Icon(OmarchyIcons.codCheck),
+                CheckboxState.unchecked => const SizedBox.shrink(),
+                CheckboxState.multiple => Container(
+                  color: foreground,
+                  margin: const EdgeInsets.all(4.0),
+                ),
+              },
             ),
-          },
-        ),
-      ),
+          ),
+        );
+      },
     );
   }
 }