+++ lib/src/widgets/navigation_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_omarchy/src/omarchy.dart';
+import 'package:flutter_omarchy/src/theme/theme.dart';
 import 'package:flutter_omarchy/src/widgets/button.dart';
 import 'package:flutter_omarchy/src/widgets/utils/title_bar.dart';
 
@@ -19,23 +19,18 @@ class OmarchyNavigationBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final omarchy = Omarchy.of(context);
+    final theme = OmarchyTheme.of(context);
 
     return Container(
-      height: omarchy.theme.text.normal.fontSize! * 4,
+      height: theme.text.normal.fontSize! * 4,
       decoration: BoxDecoration(
         border: Border(
-          bottom: BorderSide(
-            width: 1,
-            color: omarchy.theme.colors.normal.black,
-          ),
+          bottom: BorderSide(width: 1, color: theme.colors.normal.black),
         ),
       ),
       child: TitleBar(
         title: DefaultTextStyle(
-          style: omarchy.theme.text.bold.copyWith(
-            color: omarchy.theme.colors.foreground,
-          ),
+          style: theme.text.bold.copyWith(color: theme.colors.foreground),
           child: title,
         ),
         leading: leading != null