+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* Fixed localization delegates on mobile.
+* Added ProgressBar widget.
+
 ## 0.1.0
 
 * Initial release of the package.